🩹 [Patch]: Add function `Get-ScriptCommand` (#13)
## Description

This pull request introduces a new function `Get-ScriptCommand` to the
PowerShell script and adds corresponding tests to ensure its
functionality. The new function analyzes a specified PowerShell script
to extract command invocations and optionally includes call operators.

### New Functionality:

*
[`src/functions/public/Scripts/Get-ScriptCommands.ps1`](diffhunk://#diff-c59ca712a6ccbcb05a5ae179a1d551bc55f9835feb3f4d73dad9e261e3d579b9R1-R56):
Added a new function `Get-ScriptCommand` that retrieves commands used
within a specified PowerShell script, including details such as command
name, position, and file reference. It also provides an option to
include call operators in the results.

### Tests:

*
[`tests/Module.Tests.ps1`](diffhunk://#diff-a94f0e90abaaff044e3cb0327ce3a63a10cd61551269a414e4dc40ad2af8a951R45-R56):
Added a new test context for the `Get-ScriptCommands` function to verify
that it retrieves script commands correctly, ensuring the results are
not null or empty, are of the correct type, and contain expected command
names.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas